lite: colocate QAT checkpoint key canonicalization
