+++ experimental/lite/megatron/lite/model/qwen3_5/lite/checkpoint.py
@@ -13,7 +13,7 @@ import torch
 import torch.distributed as dist
 import torch.nn as nn
 from megatron.lite.model.qwen3_5.config import Qwen35Config
-from megatron.lite.model.protocol_utils import (
+from megatron.lite.primitive.quantization.qat import (
     canonical_state_key as _canonical_state_key,
 )
 from megatron.lite.primitive.ckpt.hf_weights import SafeTensorReader, unwrap_model
