+++ experimental/lite/megatron/lite/primitive/quantization/qat.py
@@ -115,6 +115,30 @@ _DEFAULT_IGNORE_PATTERNS: tuple[str, ...] = (
 )
 
 
+def canonical_state_key(key: str) -> str:
+    """Map a QAT-parametrized state key back to its logical checkpoint name.
+
+    ``torch.nn.utils.parametrize`` renames ``mod.weight`` to
+    ``mod.parametrizations.weight.original`` (the surviving BF16 master), while
+    HF checkpoints still reference the logical ``mod.weight``. Strip the
+    parametrization wrapper so loaded tensors resolve onto the master weight
+    instead of being silently dropped, which would train on random initialized
+    weights.
+
+    Only the ``.original`` master is rewritten; quantizer buffers such as
+    ``...parametrizations.weight.0.amax`` are left untouched. This is a load-side
+    mapping from an HF checkpoint into an already parametrized QAT model. Every
+    QAT-enabled model needs it regardless of whether that model can natively
+    export MXFP4; deployment export is a separate, opposite-direction contract.
+    """
+    marker = ".parametrizations."
+    if marker not in key or not key.endswith(".original"):
+        return key
+    head, rest = key.split(marker, 1)
+    attr = rest.split(".", 1)[0]
+    return f"{head}.{attr}"
+
+
 @dataclass(frozen=True)
 class QATSpec:
     """Typed, explicit opt-in QAT configuration.
@@ -738,6 +762,7 @@ __all__ = [
     "WeightFakeQuant",
     "apply_qat_to_chunks",
     "apply_qat_to_module",
+    "canonical_state_key",
     "compute_amax",
     "dequantize_weight",
     "fake_quantize_weight",
